+++ apps/renderer/src/modules/settings/hooks/use-setting-ctx.ts
@@ -2,7 +2,7 @@ import { useMemo } from "react"
 
 import { useUserRole } from "~/atoms/user"
 
-import { settings } from "../settings-glob"
+import { getMemoizedSettings } from "../settings-glob"
 import type { SettingPageContext } from "../utils"
 
 export const useSettingPageContext = (): SettingPageContext => {
@@ -12,5 +12,5 @@ export const useSettingPageContext = (): SettingPageContext => {
 
 export const useAvailableSettings = () => {
   const ctx = useSettingPageContext()
-  return useMemo(() => settings.filter((t) => !t.loader.hideIf?.(ctx)), [ctx])
+  return useMemo(() => getMemoizedSettings().filter((t) => !t.loader.hideIf?.(ctx)), [ctx])
 }
